tests/wpt/tests/svg/animations/animate-path-from-to-animation-empty-to.tentative.html

<!doctype html>
<title>SVG path 'd' from-to animation with empty 'to' value</title>
<link rel="help" href="https://svgwg.org/svg2-draft/paths.html#TheDProperty"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/SVGAnimationTestCase-testharness.js"></script>
<script src="support/animated-path-helpers.js"></script>
<svg>
  <path id="path" fill="green" d="M 40 40 L 60 40 L 60 60 L 40 60 Z">
    <animate id="animation" attributeName="d"
             from="M 0 0 L 100 0 L 100 100 L 0 100 Z" to=""
             begin="0s" dur="4s" fill="freeze"/>
  </path>
</svg>
<script>
const rootSVGElement = document.querySelector("svg");

// Non-empty from and empty to are structurally incompatible; discrete at 50%.
function sample1() {
  assert_animated_path_equals(path, "M 0 0 L 100 0 L 100 100 L 0 100 Z");
}

function sample2() {
  assert_animated_path_equals(path, "");
}

smil_async_test((t) => {
  const expectedValues = [
    ["animation", 0.001, sample1],
    ["animation", 1.999, sample1],
    ["animation", 2.001, sample2],
    ["animation", 3.999, sample2],
  ];
  runAnimationTest(t, expectedValues);
});
</script>
